Move fxconfig to better suit packaging
scripts/fxconfig.py is now just a shell, calling out the main implementation which is in fixate.core.config_util. The entry point for the utility has be listed as requried in setup.py so that it gets installed as needed.
